NUM-167 Rename package akka to api.provider
