+++ amqp-driver/src/main/scala/com/nummulus/amqp/driver/DefaultAkkaConsumer.scala
@@ -5,7 +5,7 @@ import org.slf4j.LoggerFactory
 import com.nummulus.amqp.driver.IdGenerators._
 import com.nummulus.amqp.driver.api.consumer.AmqpConsumerRequest
 import com.nummulus.amqp.driver.api.consumer.AmqpConsumerResponse
-import com.nummulus.amqp.driver.akka.AmqpRequestMessageWithProperties
+import com.nummulus.amqp.driver.api.provider.AmqpRequestMessageWithProperties
 import com.nummulus.amqp.driver.configuration.QueueConfiguration
 import com.nummulus.amqp.driver.provider.AkkaMessageConsumer
